lash-backend/pyproject.toml
[project]
name = "lash-backend"
version = "0.1.0"
description = "Backend API for Heavenly Lash Studio booking platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.6.0",
    "python-dotenv>=1.0.1",
    "pillow>=11.0.0",
]
